use portable JNDI names of beans
The old names seem to be legacy, possibly JBoss-specific paths that are
no longer supported. WebSphere used `ibm-ejb-jar-bnd.xml` to be more
specific, but that is non-portable.

During deployment, WildFly 10 logs the (allegedly portable) JNDI names
each bean is available as. Use the `app` scope, which is the narrowset
that works for these beans - module is too narrow.

See also:
- https://docs.jboss.org/author/display/WFLY10/JNDI+Reference
- https://docs.oracle.com/javaee/6/tutorial/doc/gipjf.html
- https://docs.jboss.org/author/display/WFLY10/jboss-ejb3.xml+Reference
- https://www.ibm.com/support/knowledgecenter/SSAW57_8.5.5/com.ibm.websphere.nd.doc/ae/cejb_bindingsejbfp.html