Updated wildfly deployment and Travis encoding
This commit changes how WildFly is deployed on the CD server so that
it is run as a service.  One byproduct of this is that wildfly is now
expected to sit at /opt/wildfly in addition to a relative location to the
core code base.

Travis encoded variables have been removed from the repository, and instead
should be provided via the Travis UI.  The variables needed are as follow:

CD_SERVER: the server that travis should deploy to.
CD_USER: the user with deploy permissions on the server.
CD_KEY: the private key, WRAPPED IN QUOTES, that gives the user access to the server.

Note that the CD_KEY being stored this way is a complete hack, but it seems
better than committing an encrypted key file to the repository.